requirements.txt
datasets
openpyxl
python-dotenv
py7zr
rarfile
zstandard
stanza==1.2.0
stanza-batch
nltk
indic-nlp-library
